Further sub-divided `CairoExtensions.Rendering`
